+++ gsa/src/gmp/models/report/parser.js
@@ -15,10 +15,6 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-import 'core-js/features/object/entries';
-import 'core-js/features/object/values';
-import 'core-js/features/string/includes';
-import 'core-js/features/string/starts-with';
 
 import {isDefined} from 'gmp/utils/identity';
 import {isEmpty} from 'gmp/utils/string';