+++ gsa/src/web/pages/tags/dialog.js
@@ -15,8 +15,6 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-import 'core-js/features/array/includes';
-import 'core-js/features/array/find';
 
 import React from 'react';