+++ gsa/src/web/pages/tlscertificates/dashboard/timestatusdisplay.js
@@ -15,8 +15,6 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-import 'core-js/features/object/values';
-import 'core-js/features/array/from'; // internally needed by d3-scale
 
 import {scaleOrdinal} from 'd3-scale';