+++ gsa/src/web/store/usersettings/defaultfilters/__tests__/selectors.js
@@ -1,7 +1,7 @@
 import {getUserSettingsDefaultFilter} from '../selectors';
 import Filter from 'gmp/models/filter';
 
-/* Copyright (C) 2019-2020 Greenbone Networks GmbH
+/* Copyright (C) 2019-2021 Greenbone Networks GmbH
  *
  * SPDX-License-Identifier: AGPL-3.0-or-later
  *