+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Implement policy component mutations [#2684](https://github.com/greenbone/gsa/pull/2684)
+- Refactor host component into function [#2693](https://github.com/greenbone/gsa/pull/2693)
 - Refactor Audit Component into Function [#2695](https://github.com/greenbone/gsa/pull/2695)
 - Implement policy listpage queries and mutations [#2679](https://github.com/greenbone/gsa/pull/2679)
 - Implement policy detailspage queries and mutations [#2668](https://github.com/greenbone/gsa/pull/2668)